Match structure of type constants and names
Includes moving the Flying entry in the type name table to the top of
the special types. This allows the type names table to match the type
constants ordering, which is important because the name table relies on
the defined spacing of the constants to properly assign names.
Practically, this fixes a bug that caused most physical types to be
incorrect when displayed in the pokemon's status page or pokedex entry.